Breakout KokkosExt header and partial cleanup of includes
