+++ src/details/ArborX_DetailsKokkosExtMathFunctions.hpp
@@ -1,0 +1,41 @@
+/****************************************************************************
+ * Copyright (c) 2017-2021 by the ArborX authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the ArborX library. ArborX is                       *
+ * distributed under a BSD 3-clause license. For the licensing terms see    *
+ * the LICENSE file in the top-level directory.                             *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+#ifndef ARBORX_DETAILS_KOKKOS_EXT_MATH_FUNCTIONS_HPP
+#define ARBORX_DETAILS_KOKKOS_EXT_MATH_FUNCTIONS_HPP
+
+#include <Kokkos_Macros.hpp>
+
+#include <cmath> // isfinite
+
+namespace KokkosExt
+{
+
+/** Determine whether the given floating point argument @param x has finite
+ * value.
+ *
+ * NOTE: Clang issues a warning if the std:: namespace is missing and nvcc
+ * complains about calling a __host__ function from a __host__ __device__
+ * function when it is present.
+ */
+template <typename T>
+KOKKOS_INLINE_FUNCTION bool isFinite(T x)
+{
+#ifdef __CUDA_ARCH__
+  return isfinite(x);
+#else
+  return std::isfinite(x);
+#endif
+}
+
+} // namespace KokkosExt
+
+#endif
