+++ src/details/ArborX_DetailsMortonCode.hpp
@@ -12,7 +12,7 @@
 #ifndef ARBORX_DETAILS_MORTON_CODE_UTILS_HPP
 #define ARBORX_DETAILS_MORTON_CODE_UTILS_HPP
 
-#include <ArborX_DetailsKokkosExt.hpp> // min. max
+#include <ArborX_DetailsKokkosExtMinMaxOperations.hpp> // min. max
 #include <ArborX_Exception.hpp>
 
 namespace ArborX
